+++ actionmailer/CHANGELOG.md
@@ -1,3 +1,8 @@
+*   Fix ActionMailer assertions not working when a Mail defines
+    a custom delivery job class
+
+    *Edouard Chin*
+
 *   Mails with multipart `format` blocks with implicit render now also check for
     a template name in options hash instead of only using the action name.